+++ cli/settings.py
@@ -57,161 +57,154 @@ def add_cli_options(cli_options, action):
     :param TransferAction action: action
     """
     cli_options['_action'] = action.name.lower()
-    storage_account = cli_options['storage_account']
+    storage_account = cli_options.get('storage_account')
+    azstorage = {
+        'endpoint': cli_options.get('endpoint')
+    }
     if blobxfer.util.is_not_empty(storage_account):
-        try:
-            local_resource = cli_options['local_resource']
-            if blobxfer.util.is_none_or_empty(local_resource):
-                raise KeyError()
-        except KeyError:
-            raise ValueError('--local-path must be specified')
-        try:
-            remote_path = cli_options['remote_path']
-            if blobxfer.util.is_none_or_empty(remote_path):
-                raise KeyError()
-        except KeyError:
-            raise ValueError('--remote-path must be specified')
-        # add credentials
-        try:
-            key = cli_options['access_key']
-            if blobxfer.util.is_none_or_empty(key):
-                raise KeyError()
-        except KeyError:
-            try:
-                key = cli_options['sas']
-                if blobxfer.util.is_none_or_empty(key):
-                    raise KeyError()
-            except KeyError:
-                raise RuntimeError('access key or sas must be provided')
-        azstorage = {
-            'endpoint': cli_options['endpoint'],
-            'accounts': {
-                storage_account: key
-            }
+        azstorage['accounts'] = {
+            storage_account: (
+                cli_options.get('access_key') or cli_options.get('sas')
+            )
         }
-        del key
-        # construct "argument" from cli options
-        sa_rp = {storage_account: remote_path}
-        if action == TransferAction.Upload:
-            arg = {
-                'source': [local_resource],
-                'destination': [sa_rp],
-                'include': cli_options['include'],
-                'exclude': cli_options['exclude'],
-                'options': {
-                    'chunk_size_bytes': cli_options['chunk_size_bytes'],
-                    'delete_extraneous_destination': cli_options['delete'],
-                    'mode': cli_options['mode'],
-                    'one_shot_bytes': cli_options['one_shot_bytes'],
-                    'overwrite': cli_options['overwrite'],
-                    'recursive': cli_options['recursive'],
-                    'rename': cli_options['rename'],
-                    'rsa_private_key': cli_options['rsa_private_key'],
-                    'rsa_private_key_passphrase': cli_options[
-                        'rsa_private_key_passphrase'],
-                    'rsa_public_key': cli_options['rsa_public_key'],
-                    'skip_on': {
-                        'filesize_match': cli_options[
-                            'skip_on_filesize_match'],
-                        'lmt_ge': cli_options['skip_on_lmt_ge'],
-                        'md5_match': cli_options['skip_on_md5_match'],
-                    },
-                    'store_file_properties': {
-                        'attributes': cli_options['file_attributes'],
-                        'md5': cli_options['file_md5'],
-                    },
-                    'strip_components': cli_options['strip_components'],
-                    'vectored_io': {
-                        'stripe_chunk_size_bytes': cli_options[
-                            'stripe_chunk_size_bytes'],
-                        'distribution_mode': cli_options['distribution_mode'],
-                    },
+    sa_rp = {
+        storage_account: cli_options.get('remote_path')
+    }
+    local_resource = cli_options.get('local_resource')
+    # construct "argument" from cli options
+    if action == TransferAction.Upload:
+        arg = {
+            'source': [local_resource] if local_resource is not None else None,
+            'destination': [sa_rp] if sa_rp[storage_account] is not None else
+            None,
+            'include': cli_options.get('include'),
+            'exclude': cli_options.get('exclude'),
+            'options': {
+                'chunk_size_bytes': cli_options.get('chunk_size_bytes'),
+                'delete_extraneous_destination': cli_options.get('delete'),
+                'mode': cli_options.get('mode'),
+                'one_shot_bytes': cli_options.get('one_shot_bytes'),
+                'overwrite': cli_options.get('overwrite'),
+                'recursive': cli_options.get('recursive'),
+                'rename': cli_options.get('rename'),
+                'rsa_private_key': cli_options.get('rsa_private_key'),
+                'rsa_private_key_passphrase': cli_options.get(
+                    'rsa_private_key_passphrase'),
+                'rsa_public_key': cli_options.get('rsa_public_key'),
+                'skip_on': {
+                    'filesize_match': cli_options.get(
+                        'skip_on_filesize_match'),
+                    'lmt_ge': cli_options.get('skip_on_lmt_ge'),
+                    'md5_match': cli_options.get('skip_on_md5_match'),
                 },
-            }
-        elif action == TransferAction.Download:
-            arg = {
-                'source': [sa_rp],
-                'destination': local_resource,
-                'include': cli_options['include'],
-                'exclude': cli_options['exclude'],
-                'options': {
-                    'check_file_md5': cli_options['file_md5'],
-                    'chunk_size_bytes': cli_options['chunk_size_bytes'],
-                    'delete_extraneous_destination': cli_options['delete'],
-                    'mode': cli_options['mode'],
-                    'overwrite': cli_options['overwrite'],
-                    'recursive': cli_options['recursive'],
-                    'rename': cli_options['rename'],
-                    'rsa_private_key': cli_options['rsa_private_key'],
-                    'rsa_private_key_passphrase': cli_options[
-                        'rsa_private_key_passphrase'],
-                    'restore_file_attributes': cli_options['file_attributes'],
-                    'skip_on': {
-                        'filesize_match': cli_options[
-                            'skip_on_filesize_match'],
-                        'lmt_ge': cli_options['skip_on_lmt_ge'],
-                        'md5_match': cli_options['skip_on_md5_match'],
-                    },
+                'store_file_properties': {
+                    'attributes': cli_options.get('file_attributes'),
+                    'md5': cli_options.get('file_md5'),
                 },
-            }
-        elif action == TransferAction.Synccopy:
-            try:
-                sync_copy_dest_storage_account = \
-                    cli_options['sync_copy_dest_storage_account']
-                if blobxfer.util.is_none_or_empty(
-                        sync_copy_dest_storage_account):
-                    raise KeyError()
-            except KeyError:
-                raise ValueError(
-                    '--sync-copy-dest-storage-account must be specified')
-            try:
-                sync_copy_dest_remote_path = \
-                    cli_options['sync_copy_dest_remote_path']
-                if blobxfer.util.is_none_or_empty(sync_copy_dest_remote_path):
-                    raise KeyError()
-            except KeyError:
-                raise ValueError(
-                    '--sync-copy-dest-remote-path must be specified')
-            arg = {
-                'source': sa_rp,
-                'destination': [
-                    {
-                        sync_copy_dest_storage_account:
-                        sync_copy_dest_remote_path
-                    }
-                ],
-                'include': cli_options['include'],
-                'exclude': cli_options['exclude'],
-                'options': {
-                    'chunk_size_bytes': cli_options['chunk_size_bytes'],
-                    'dest_mode': cli_options['sync_copy_dest_mode'],
-                    'mode': cli_options['mode'],
-                    'overwrite': cli_options['overwrite'],
-                    'skip_on': {
-                        'filesize_match': cli_options[
-                            'skip_on_filesize_match'],
-                        'lmt_ge': cli_options['skip_on_lmt_ge'],
-                        'md5_match': cli_options['skip_on_md5_match'],
-                    },
+                'strip_components': cli_options.get('strip_components'),
+                'vectored_io': {
+                    'stripe_chunk_size_bytes': cli_options.get(
+                        'stripe_chunk_size_bytes'),
+                    'distribution_mode': cli_options.get('distribution_mode'),
                 },
-            }
-            try:
-                destkey = cli_options['sync_copy_dest_access_key']
-                if blobxfer.util.is_none_or_empty(destkey):
-                    raise KeyError()
-            except KeyError:
-                try:
-                    destkey = cli_options['sync_copy_dest_sas']
-                    if blobxfer.util.is_none_or_empty(destkey):
-                        raise KeyError()
-                except KeyError:
-                    raise RuntimeError(
-                        'destination access key or sas must be provided')
-            azstorage['accounts'][
-                cli_options['sync_copy_dest_storage_account']] = destkey
-            del destkey
-        cli_options[action.name.lower()] = arg
+            },
+        }
+    elif action == TransferAction.Download:
+        arg = {
+            'source': [sa_rp] if sa_rp[storage_account] is not None else None,
+            'destination': local_resource if local_resource is not None else
+            None,
+            'include': cli_options.get('include'),
+            'exclude': cli_options.get('exclude'),
+            'options': {
+                'check_file_md5': cli_options.get('file_md5'),
+                'chunk_size_bytes': cli_options.get('chunk_size_bytes'),
+                'delete_extraneous_destination': cli_options.get('delete'),
+                'mode': cli_options.get('mode'),
+                'overwrite': cli_options.get('overwrite'),
+                'recursive': cli_options.get('recursive'),
+                'rename': cli_options.get('rename'),
+                'rsa_private_key': cli_options.get('rsa_private_key'),
+                'rsa_private_key_passphrase': cli_options.get(
+                    'rsa_private_key_passphrase'),
+                'restore_file_attributes': cli_options.get(
+                    'file_attributes'),
+                'skip_on': {
+                    'filesize_match': cli_options.get(
+                        'skip_on_filesize_match'),
+                    'lmt_ge': cli_options.get('skip_on_lmt_ge'),
+                    'md5_match': cli_options.get('skip_on_md5_match'),
+                },
+            },
+        }
+    elif action == TransferAction.Synccopy:
+        sync_copy_dest_storage_account = cli_options.get(
+            'sync_copy_dest_storage_account')
+        sync_copy_dest_remote_path = cli_options.get(
+            'sync_copy_dest_remote_path')
+        if (sync_copy_dest_storage_account is not None and
+                sync_copy_dest_remote_path is not None):
+            sync_copy_dest = [
+                {
+                    sync_copy_dest_storage_account:
+                    sync_copy_dest_remote_path
+                }
+            ]
+            azstorage['accounts'][sync_copy_dest_storage_account] = (
+                cli_options.get('sync_copy_dest_access_key') or
+                cli_options.get('sync_copy_dest_sas')
+            )
+        else:
+            sync_copy_dest = None
+        arg = {
+            'source': sa_rp if sa_rp[storage_account] is not None else None,
+            'destination': sync_copy_dest,
+            'include': cli_options.get('include'),
+            'exclude': cli_options.get('exclude'),
+            'options': {
+                'chunk_size_bytes': cli_options.get('chunk_size_bytes'),
+                'dest_mode': cli_options.get('sync_copy_dest_mode'),
+                'mode': cli_options.get('mode'),
+                'overwrite': cli_options.get('overwrite'),
+                'skip_on': {
+                    'filesize_match': cli_options.get(
+                        'skip_on_filesize_match'),
+                    'lmt_ge': cli_options.get('skip_on_lmt_ge'),
+                    'md5_match': cli_options.get('skip_on_md5_match'),
+                },
+            },
+        }
+    count = 0
+    if arg['source'] is None:
+        arg.pop('source')
+        count += 1
+    if arg['destination'] is None:
+        arg.pop('destination')
+        count += 1
+    if count == 1:
+        raise ValueError(
+            '--local-path and --remote-path must be specified together '
+            'through the commandline')
+    if 'accounts' in azstorage:
         cli_options['azure_storage'] = azstorage
+    cli_options[action.name.lower()] = arg
+
+
+def _merge_setting(cli_options, conf, name, name_cli=None, default=None):
+    # type: (dict, dict, str, str, Any) -> Any
+    """Merge a setting, preferring the CLI option if set
+    :param dict cli_options: cli options
+    :param dict conf: configuration sub-block
+    :param str name: key name
+    :param str name_cli: override key name from cli_options
+    :param Any default: default value to set if missing
+    :rtype: Any
+    :return: merged setting value
+    """
+    val = cli_options.get(name_cli or name)
+    if val is None:
+        val = conf.get(name, default)
+    return val
 
 
 def merge_settings(config, cli_options):
@@ -225,37 +218,59 @@ def merge_settings(config, cli_options):
             action != TransferAction.Download.name.lower() and
             action != TransferAction.Synccopy.name.lower()):
         raise ValueError('invalid action: {}'.format(action))
-    # create action options
-    if action not in config:
-        config[action] = []
-    # merge any argument options
-    if action in cli_options:
-        config[action].append(cli_options[action])
     # merge credentials
     if 'azure_storage' in cli_options:
         if 'azure_storage' not in config:
             config['azure_storage'] = {}
         config['azure_storage'] = blobxfer.util.merge_dict(
             config['azure_storage'], cli_options['azure_storage'])
-    # merge general options
+    if blobxfer.util.is_none_or_empty(config['azure_storage']):
+        raise ValueError('azure storage settings not specified')
+    # create action options
+    if action not in config:
+        config[action] = []
+    # append full specs, if they exist
+    if action in cli_options:
+        if 'source' in cli_options[action]:
+            srcdst = {
+                'source': cli_options[action]['source'],
+                'destination': cli_options[action]['destination'],
+            }
+            cli_options[action].pop('source')
+            cli_options[action].pop('destination')
+            config[action].append(srcdst)
+    # merge general and concurrency options
     if 'options' not in config:
         config['options'] = {}
-    config['options']['log_file'] = cli_options['log_file']
-    config['options']['progress_bar'] = cli_options['progress_bar']
-    config['options']['resume_file'] = cli_options['resume_file']
-    config['options']['timeout_sec'] = cli_options['timeout']
-    config['options']['verbose'] = cli_options['verbose']
-    # merge concurrency options
     if 'concurrency' not in config['options']:
         config['options']['concurrency'] = {}
-    config['options']['concurrency']['crypto_processes'] = \
-        cli_options['crypto_processes']
-    config['options']['concurrency']['disk_threads'] = \
-        cli_options['disk_threads']
-    config['options']['concurrency']['md5_processes'] = \
-        cli_options['md5_processes']
-    config['options']['concurrency']['transfer_threads'] = \
-        cli_options['transfer_threads']
+    options = {
+        'log_file': _merge_setting(cli_options, config['options'], 'log_file'),
+        'progress_bar': _merge_setting(
+            cli_options, config['options'], 'progress_bar', default=True),
+        'resume_file': _merge_setting(
+            cli_options, config['options'], 'resume_file'),
+        'timeout_sec': _merge_setting(
+            cli_options, config['options'], 'timeout'),
+        'verbose': _merge_setting(
+            cli_options, config['options'], 'verbose', default=False),
+        'concurrency': {
+            'crypto_processes': _merge_setting(
+                cli_options, config['options']['concurrency'],
+                'crypto_processes', default=0),
+            'disk_threads': _merge_setting(
+                cli_options, config['options']['concurrency'],
+                'disk_threads', default=0),
+            'md5_processes': _merge_setting(
+                cli_options, config['options']['concurrency'],
+                'md5_processes', default=0),
+            'transfer_threads': _merge_setting(
+                cli_options, config['options']['concurrency'],
+                'transfer_threads', default=0),
+        }
+    }
+    config['options'] = options
+    cli_options = cli_options[action]
 
 
 def create_azure_storage_credentials(config, general_options):
@@ -268,7 +283,7 @@ def create_azure_storage_credentials(config, general_options):
     :return: credentials object
     """
     creds = blobxfer.operations.azure.StorageCredentials(general_options)
-    endpoint = config['azure_storage']['endpoint']
+    endpoint = config['azure_storage'].get('endpoint', 'core.windows.net')
     for name in config['azure_storage']['accounts']:
         key = config['azure_storage']['accounts'][name]
         creds.add_storage_account(name, key, endpoint)
@@ -283,26 +298,27 @@ def create_general_options(config, action):
     :rtype: blobxfer.models.options.General
     :return: general options object
     """
-    conc = config['options'].get('concurrency', {})
+    conc = config['options']['concurrency']
     return blobxfer.models.options.General(
         concurrency=blobxfer.models.options.Concurrency(
-            crypto_processes=conc.get('crypto_processes', 0),
-            disk_threads=conc.get('disk_threads', 0),
-            md5_processes=conc.get('md5_processes', 0),
-            transfer_threads=conc.get('transfer_threads', 0),
+            crypto_processes=conc['crypto_processes'],
+            disk_threads=conc['disk_threads'],
+            md5_processes=conc['md5_processes'],
+            transfer_threads=conc['transfer_threads'],
             action=action.value[0],
         ),
-        log_file=config['options'].get('log_file', None),
-        progress_bar=config['options'].get('progress_bar', True),
-        resume_file=config['options'].get('resume_file', None),
-        timeout_sec=config['options'].get('timeout_sec', None),
-        verbose=config['options'].get('verbose', False),
+        log_file=config['options']['log_file'],
+        progress_bar=config['options']['progress_bar'],
+        resume_file=config['options']['resume_file'],
+        timeout_sec=config['options']['timeout_sec'],
+        verbose=config['options']['verbose'],
     )
 
 
-def create_download_specifications(config):
-    # type: (dict) -> List[blobxfer.models.download.Specification]
+def create_download_specifications(cli_options, config):
+    # type: (dict, dict) -> List[blobxfer.models.download.Specification]
     """Create a list of Download Specification objects from configuration
+    :param dict cli_options: cli options
     :param dict config: config dict
     :rtype: list
     :return: list of Download Specification objects
@@ -310,7 +326,8 @@ def create_download_specifications(config):
     specs = []
     for conf in config['download']:
         # create download options
-        mode = conf['options'].get('mode', 'auto').lower()
+        mode = _merge_setting(
+            cli_options, conf['options'], 'mode', default='auto').lower()
         if mode == 'auto':
             mode = blobxfer.models.azure.StorageModes.Auto
         elif mode == 'append':
@@ -324,9 +341,12 @@ def create_download_specifications(config):
         else:
             raise ValueError('unknown mode: {}'.format(mode))
         # load RSA private key PEM file if specified
-        rpk = conf['options'].get('rsa_private_key', None)
+        rpk = _merge_setting(
+            cli_options, conf['options'], 'rsa_private_key', default=None)
         if blobxfer.util.is_not_empty(rpk):
-            rpkp = conf['options'].get('rsa_private_key_passphrase', None)
+            rpkp = _merge_setting(
+                cli_options, conf['options'], 'rsa_private_key_passphrase',
+                default=None)
             rpk = blobxfer.operations.crypto.load_rsa_private_key_file(
                 rpk, rpkp)
         else:
@@ -335,22 +355,37 @@ def create_download_specifications(config):
         sod = conf['options'].get('skip_on', {})
         ds = blobxfer.models.download.Specification(
             download_options=blobxfer.models.options.Download(
-                check_file_md5=conf['options'].get('check_file_md5', False),
-                chunk_size_bytes=conf['options'].get('chunk_size_bytes', 0),
-                delete_extraneous_destination=conf['options'].get(
-                    'delete_extraneous_destination', False),
+                check_file_md5=_merge_setting(
+                    cli_options, conf['options'], 'check_file_md5',
+                    default=False),
+                chunk_size_bytes=_merge_setting(
+                    cli_options, conf['options'], 'chunk_size_bytes',
+                    default=0),
+                delete_extraneous_destination=_merge_setting(
+                    cli_options, conf['options'],
+                    'delete_extraneous_destination', default=False),
                 mode=mode,
-                overwrite=conf['options'].get('overwrite', True),
-                recursive=conf['options'].get('recursive', True),
-                rename=conf['options'].get('rename', False),
-                restore_file_attributes=conf[
-                    'options'].get('restore_file_attributes', False),
+                overwrite=_merge_setting(
+                    cli_options, conf['options'], 'overwrite', default=True),
+                recursive=_merge_setting(
+                    cli_options, conf['options'], 'recursive', default=True),
+                rename=_merge_setting(
+                    cli_options, conf['options'], 'rename', default=False),
+                restore_file_attributes=_merge_setting(
+                    cli_options, conf['options'], 'restore_file_attributes',
+                    default=False),
                 rsa_private_key=rpk,
             ),
             skip_on_options=blobxfer.models.options.SkipOn(
-                filesize_match=sod.get('filesize_match', False),
-                lmt_ge=sod.get('lmt_ge', False),
-                md5_match=sod.get('md5_match', False),
+                filesize_match=_merge_setting(
+                    cli_options, sod, 'filesize_match',
+                    name_cli='skip_on_filesize_match', default=False),
+                lmt_ge=_merge_setting(
+                    cli_options, sod, 'lmt_ge', name_cli='skip_on_lmt_ge',
+                    default=False),
+                md5_match=_merge_setting(
+                    cli_options, sod, 'md5_match',
+                    name_cli='skip_on_md5_match', default=False),
             ),
             local_destination_path=blobxfer.models.download.
             LocalDestinationPath(
@@ -365,10 +400,10 @@ def create_download_specifications(config):
             sa = next(iter(src))
             asp = blobxfer.operations.azure.SourcePath()
             asp.add_path_with_storage_account(src[sa], sa)
-            incl = conf.get('include', None)
+            incl = _merge_setting(cli_options, conf, 'include', default=None)
             if blobxfer.util.is_not_empty(incl):
                 asp.add_includes(incl)
-            excl = conf.get('exclude', None)
+            excl = _merge_setting(cli_options, conf, 'exclude', default=None)
             if blobxfer.util.is_not_empty(excl):
                 asp.add_excludes(excl)
             ds.add_azure_source_path(asp)
@@ -377,9 +412,10 @@ def create_download_specifications(config):
     return specs
 
 
-def create_synccopy_specifications(config):
-    # type: (dict) -> List[blobxfer.models.synccopy.Specification]
+def create_synccopy_specifications(cli_options, config):
+    # type: (dict, dict) -> List[blobxfer.models.synccopy.Specification]
     """Create a list of SyncCopy Specification objects from configuration
+    :param dict cli_options: cli options
     :param dict config: config dict
     :rtype: list
     :return: list of SyncCopy Specification objects
@@ -387,7 +423,8 @@ def create_synccopy_specifications(config):
     specs = []
     for conf in config['synccopy']:
         # get source mode
-        mode = conf['options'].get('mode', 'auto').lower()
+        mode = _merge_setting(
+            cli_options, conf['options'], 'mode', default='auto').lower()
         if mode == 'auto':
             mode = blobxfer.models.azure.StorageModes.Auto
         elif mode == 'append':
@@ -401,7 +438,9 @@ def create_synccopy_specifications(config):
         else:
             raise ValueError('unknown source mode: {}'.format(mode))
         # get destination mode
-        destmode = conf['options'].get('dest_mode')
+        destmode = _merge_setting(
+            cli_options, conf['options'], 'dest_mode',
+            name_cli='sync_copy_dest_mode')
         if blobxfer.util.is_none_or_empty(destmode):
             destmode = mode
         else:
@@ -422,17 +461,26 @@ def create_synccopy_specifications(config):
         sod = conf['options'].get('skip_on', {})
         scs = blobxfer.models.synccopy.Specification(
             synccopy_options=blobxfer.models.options.SyncCopy(
-                delete_extraneous_destination=conf['options'].get(
-                    'delete_extraneous_destination', False),
+                delete_extraneous_destination=_merge_setting(
+                    cli_options, conf['options'],
+                    'delete_extraneous_destination', default=False),
                 dest_mode=destmode,
                 mode=mode,
-                overwrite=conf['options'].get('overwrite', True),
-                recursive=conf['options'].get('recursive', True),
+                overwrite=_merge_setting(
+                    cli_options, conf['options'], 'overwrite', default=True),
+                recursive=_merge_setting(
+                    cli_options, conf['options'], 'recursive', default=True),
             ),
             skip_on_options=blobxfer.models.options.SkipOn(
-                filesize_match=sod.get('filesize_match', False),
-                lmt_ge=sod.get('lmt_ge', False),
-                md5_match=sod.get('md5_match', False),
+                filesize_match=_merge_setting(
+                    cli_options, sod, 'filesize_match',
+                    name_cli='skip_on_filesize_match', default=False),
+                lmt_ge=_merge_setting(
+                    cli_options, sod, 'lmt_ge', name_cli='skip_on_lmt_ge',
+                    default=False),
+                md5_match=_merge_setting(
+                    cli_options, sod, 'md5_match',
+                    name_cli='skip_on_md5_match', default=False),
             ),
         )
         # create remote source paths
@@ -440,10 +488,10 @@ def create_synccopy_specifications(config):
             sa = next(iter(src))
             asp = blobxfer.operations.azure.SourcePath()
             asp.add_path_with_storage_account(src[sa], sa)
-            incl = conf.get('include', None)
+            incl = _merge_setting(cli_options, conf, 'include', default=None)
             if blobxfer.util.is_not_empty(incl):
                 asp.add_includes(incl)
-            excl = conf.get('exclude', None)
+            excl = _merge_setting(cli_options, conf, 'exclude', default=None)
             if blobxfer.util.is_not_empty(excl):
                 asp.add_excludes(excl)
             scs.add_azure_source_path(asp)
@@ -461,17 +509,21 @@ def create_synccopy_specifications(config):
     return specs
 
 
-def create_upload_specifications(config):
-    # type: (dict) -> List[blobxfer.models.upload.Specification]
+def create_upload_specifications(cli_options, config):
+    # type: (dict, dict) -> List[blobxfer.models.upload.Specification]
     """Create a list of Upload Specification objects from configuration
+    :param dict cli_options: cli options
     :param dict config: config dict
     :rtype: list
     :return: list of Upload Specification objects
     """
     specs = []
     for conf in config['upload']:
+        if 'options' not in conf:
+            conf['options'] = {}
         # create upload options
-        mode = conf['options'].get('mode', 'auto').lower()
+        mode = _merge_setting(
+            cli_options, conf['options'], 'mode', default='auto').lower()
         if mode == 'auto':
             mode = blobxfer.models.azure.StorageModes.Auto
         elif mode == 'append':
@@ -485,14 +537,16 @@ def create_upload_specifications(config):
         else:
             raise ValueError('unknown mode: {}'.format(mode))
         # load RSA public key PEM if specified
-        rpk = conf['options'].get('rsa_public_key', None)
+        rpk = _merge_setting(cli_options, conf['options'], 'rsa_public_key')
         if blobxfer.util.is_not_empty(rpk):
             rpk = blobxfer.operations.crypto.load_rsa_public_key_file(rpk)
         if rpk is None:
             # load RSA private key PEM file if specified
-            rpk = conf['options'].get('rsa_private_key', None)
+            rpk = _merge_setting(
+                cli_options, conf['options'], 'rsa_private_key')
             if blobxfer.util.is_not_empty(rpk):
-                rpkp = conf['options'].get('rsa_private_key_passphrase', None)
+                rpkp = _merge_setting(
+                    cli_options, conf['options'], 'rsa_private_key_passphrase')
                 rpk = blobxfer.operations.crypto.load_rsa_private_key_file(
                     rpk, rpkp)
                 rpk = rpk.public_key()
@@ -501,10 +555,10 @@ def create_upload_specifications(config):
         # create local source paths
         lsp = blobxfer.models.upload.LocalSourcePath()
         lsp.add_paths(conf['source'])
-        incl = conf.get('include', None)
+        incl = _merge_setting(cli_options, conf, 'include', default=None)
         if blobxfer.util.is_not_empty(incl):
             lsp.add_includes(incl)
-        excl = conf.get('exclude', None)
+        excl = _merge_setting(cli_options, conf, 'exclude', default=None)
         if blobxfer.util.is_not_empty(excl):
             lsp.add_excludes(excl)
         # create specification
@@ -513,32 +567,54 @@ def create_upload_specifications(config):
         sod = conf['options'].get('skip_on', {})
         us = blobxfer.models.upload.Specification(
             upload_options=blobxfer.models.options.Upload(
-                chunk_size_bytes=conf['options'].get('chunk_size_bytes', 0),
-                delete_extraneous_destination=conf['options'].get(
-                    'delete_extraneous_destination', False),
+                chunk_size_bytes=_merge_setting(
+                    cli_options, conf['options'], 'chunk_size_bytes',
+                    default=0),
+                delete_extraneous_destination=_merge_setting(
+                    cli_options, conf['options'],
+                    'delete_extraneous_destination', default=False),
                 mode=mode,
-                one_shot_bytes=conf['options'].get('one_shot_bytes', 0),
-                overwrite=conf['options'].get('overwrite', True),
-                recursive=conf['options'].get('recursive', True),
-                rename=conf['options'].get('rename', False),
+                one_shot_bytes=_merge_setting(
+                    cli_options, conf['options'], 'one_shot_bytes', default=0),
+                overwrite=_merge_setting(
+                    cli_options, conf['options'], 'overwrite', default=True),
+                recursive=_merge_setting(
+                    cli_options, conf['options'], 'recursive', default=True),
+                rename=_merge_setting(
+                    cli_options, conf['options'], 'rename', default=False),
                 rsa_public_key=rpk,
                 store_file_properties=blobxfer.models.options.FileProperties(
-                    attributes=sfp.get('attributes', False),
-                    md5=sfp.get('md5', False),
+                    attributes=_merge_setting(
+                        cli_options, sfp, 'attributes',
+                        name_cli='file_attributes', default=False),
+                    md5=_merge_setting(
+                        cli_options, sfp, 'md5', name_cli='file_md5',
+                        default=False),
                 ),
-                strip_components=conf['options'].get('strip_components', 1),
+                strip_components=_merge_setting(
+                    cli_options, conf['options'], 'strip_components',
+                    default=0),
                 vectored_io=blobxfer.models.options.VectoredIo(
-                    stripe_chunk_size_bytes=vio.get(
-                        'stripe_chunk_size_bytes', 1073741824),
+                    stripe_chunk_size_bytes=_merge_setting(
+                        cli_options, vio, 'stripe_chunk_size_bytes',
+                        default=1073741824),
                     distribution_mode=blobxfer.
                     models.upload.VectoredIoDistributionMode(
-                        vio.get('distribution_mode', 'disabled').lower()),
+                        _merge_setting(
+                            cli_options, vio, 'distribution_mode',
+                            default='disabled').lower()),
                 ),
             ),
             skip_on_options=blobxfer.models.options.SkipOn(
-                filesize_match=sod.get('filesize_match', False),
-                lmt_ge=sod.get('lmt_ge', False),
-                md5_match=sod.get('md5_match', False),
+                filesize_match=_merge_setting(
+                    cli_options, sod, 'filesize_match',
+                    name_cli='skip_on_filesize_match', default=False),
+                lmt_ge=_merge_setting(
+                    cli_options, sod, 'lmt_ge', name_cli='skip_on_lmt_ge',
+                    default=False),
+                md5_match=_merge_setting(
+                    cli_options, sod, 'md5_match',
+                    name_cli='skip_on_md5_match', default=False),
             ),
             local_source_path=lsp,
         )